bon je viens de faire encore quelque modification parce que il y'avait quelque petite incorence donc la maintenant je pense que tout est bon le (nb: logiquement pour test utilise juste les controller tout est deja labas ) pas la peine de toucher au reste du code sinon tu vas rencontrer des problème tout est lié ensemble donc attention (utilise le controller reservation pour essayer de faire des reservation , annulerreservation , tu peux aussi recuperer tout les reservation pour les afficher chez toi et aussi fait la meme chose pour evenement , forum, groupe et aussi etudiant tu as juste un controller ajouterEtudiant c'est fais exprès demain je vais ajouter le reste tu peux juste tester ça pour le moment )

+++ src/main/java/com/dembaandousmane/gest_priere/groupe/repository/GroupeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroupeRepository extends JpaRepository<Groupe, Integer> {
+public interface GroupeRepository extends JpaRepository<Groupe, Long> {
 }
